backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "fastapi>=0.128.0",
    "google-re2>=1.1.20251105",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-magic>=0.4.27",
    "python-multipart>=0.0.22",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.46",
    "torch>=2.10.0",
    "torchaudio>=2.10.0",
    "transformers>=5.0.0",
    "uvicorn>=0.40.0",
]

[project.optional-dependencies]
dev = [
    "black>=26.1.0",
    "httpx>=0.28.1",
    "ipykernel>=7.1.0",
    "isort>=7.0.0",
    "pre-commit>=4.5.1",
    "pydocstyle>=6.3.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.14",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
# use uv sync --extra dev
# uv add <package_name> --optional dev
